+++ js/data.js
@@ -1,0 +1,51 @@
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Алексей',
+  'Екатерина',
+  'Иван',
+  'Николай',
+  'Татьяна',
+  'Сергей',
+  'Варвара',
+  'Леонид',
+  'Виктория',
+  'Анна',
+];
+
+const DESCRIPTIONS = [
+  'Полевые маки в ярких оттенках',
+  'Тропический пляж с белым песком',
+  'Природный водопад с пением воды',
+  'Городская скульптура на фоне неба',
+  'Солнечный закат на горизонте',
+  'Старинная архитектура в историческом центре',
+  'Заснеженные вершины гор',
+  'Разноцветные лодки на озере',
+  'Взлетающий шар в небе',
+  'Причудливые облака на закате',
+  'Водопад с облачным дождем',
+  'Закат на горном хребте',
+  'Цветущие сакуры в парке',
+  'Мощный вулкан с извержением',
+  'Магическая полярная сияние',
+  'Раскаты грома над океаном',
+  'Пасторальная деревня с зелеными полями',
+  'Грандиозный фейерверк ночью',
+  'Загадочный лес с пышной растительностью',
+  'Искусственное озеро с парящими лебедями',
+  'Солнечный закат над пустыней',
+  'Расцветающий сад с яркими цветами',
+  'Кристально чистая горная река',
+  'Архитектурное чудо в историческом центре',
+  'Панорамный вид на океан'
+];
+
+export {COMMENTS, NAMES, DESCRIPTIONS};
